Split public template tests into a folder and drop its max-lines exemption
The public-page template test file had grown to 1,909 lines and was
grandfathered into biome's noExcessiveLinesPerFile exemption list. Split
it into test/templates/public/ — grouped by what each part renders
(single-listing ticket pages, OG/status pages, the main ticket page's
package and field/form halves, listing date/location and package grouping,
the ticket-view page, the day-count selector, and listing images) — so
every file sits under 400 lines, then removed the file's exemption entry.

The move surfaced repeated ticketPage render setup, now shared helpers in
helpers.ts: ticketListing, singleListingPageHtml, bigAndSmallListings,
evenSplitPackages, listingB, and registerPublicTemplateHooks (the shared
beforeAll/afterEach). All original tests carry over unchanged and pass;
jscpd is back to 0% duplication.
